Modularize app.py into callbacks, selectors, and UI elements
