+++ src/callbacks/scatterplot_callback.py
@@ -1,0 +1,52 @@
+from dash import callback, Output, Input
+import altair as alt
+from data.data_loader import load_data
+
+data = load_data()
+
+# Scatterplot Callback
+@callback(
+    Output('scatterplot', 'spec'),
+    Input('year_dropdown', 'value'),
+    Input('duration_min', 'value'),
+    Input('duration_max', 'value'),
+    Input('bpm_range', 'value'),
+    Input('genre_dropdown', 'value')
+)
+def update_scatterplot(selected_year, selected_duration_min, selected_duration_max, selected_bpm_range, selected_genres):
+    filtered_data = data[
+        (data['year'] == selected_year) &
+        (data['duration'] >= (selected_duration_min if selected_duration_min is not None else 0)) &
+        (data['duration'] <= (selected_duration_max if selected_duration_max is not None else data['duration'].max())) &
+        (data['bpm'] >= selected_bpm_range[0]) &
+        (data['bpm'] <= selected_bpm_range[1])
+    ]
+
+    if selected_genres:
+        filtered_data = filtered_data[filtered_data['genre'].isin(selected_genres)]
+
+    return (
+        alt.Chart(
+            filtered_data, width='container'
+        ).mark_point(filled=True, size=100).encode(
+            x=alt.X('popularity', title='Popularity', scale=alt.Scale(zero=False)),
+            y=alt.Y('danceability', title='Danceability', scale=alt.Scale(zero=False)),
+            tooltip=[
+                alt.Tooltip('title', title='Song Title'), 
+                alt.Tooltip('artist', title='Artist'), 
+                alt.Tooltip('popularity', title='Popularity'), 
+                alt.Tooltip('danceability', title='Danceability'),
+                alt.Tooltip('duration', title='Duration (Seconds)'),
+                alt.Tooltip('bpm', title='BPM')
+            ]
+        ).properties(
+            title=f'Danceability vs. Popularity in {selected_year}'
+        ).configure_axis(
+            labelFontSize=20,
+            titleFontSize=24
+        ).configure_title(
+            fontSize=24
+        ).configure(
+            background='#282828'
+        ).to_dict()
+    )
