+++ src/components/artist_list.py
@@ -1,0 +1,4 @@
+from dash import html
+
+# Top artists chart
+artist_elements = html.Div(id='artist_list')
